feat/Client : refactoring code for add subcategories for accountant
