+++ client/src/components/displayCategorySubcategory/DisplayCategorySubcategory.tsx
@@ -1,7 +1,6 @@
 import {
   useGetCategoriesQuery,
   useUpdateSubcategoryMutation,
-  useAddSubcategoryMutation,
 } from "../../types/graphql-types";
 import { useState } from "react";
 import useNotification from "../../hooks/useNotification";
@@ -24,7 +23,6 @@ import {
 import KeyboardArrowDownIcon from "@mui/icons-material/KeyboardArrowDown";
 import KeyboardArrowUpIcon from "@mui/icons-material/KeyboardArrowUp";
 import CreateIcon from "@mui/icons-material/Create";
-import AddCircleOutline from "@mui/icons-material/AddCircleOutline";
 import ArrowForwardIosIcon from "@mui/icons-material/ArrowForwardIos";
 
 import { Button } from "@mui/material";
@@ -50,6 +48,7 @@ function createData(
 }
 
 import UpdateCategory from "../updateCategory/UpdateCategory";
+import AddSubcategory from "../addSubcategory/AddSubcategory";
 
 function Row(props: {
   row: ReturnType<typeof createData>;
@@ -58,68 +57,18 @@ function Row(props: {
 }) {
   const { row, rows, creditDebitOptions } = props;
   const [open, setOpen] = useState(false);
-  const [showInput, setShowInput] = useState(false);
   const [editCategory, setEditCategory] = useState(false);
 
   const [editSubcategoryId, setEditSubcategoryId] = useState<number>(0);
 
-  const [newSubcategoryLabel, setNewSubcategoryLabel] = useState("");
-  const [newSubcategoryCode, setNewSubcategoryCode] = useState("");
-
   const [updateSubcategoryLabel, setUpdateSubcategoryLabel] = useState("");
   const [updateSubcategoryCode, setUpdateSubcategoryCode] = useState("");
-
-  const [addASubcategory] = useAddSubcategoryMutation();
 
   const [updateSubcategory] = useUpdateSubcategoryMutation();
 
   const { notifySuccess, notifyError } = useNotification();
 
   const theme = useTheme();
-
-  const handleValidation = (): boolean => {
-    if (!newSubcategoryLabel && !newSubcategoryCode) {
-      notifyError("Veuillez remplir les champs");
-      return false;
-    }
-    if (!newSubcategoryLabel) {
-      notifyError("Veuillez remplir le champ 'Nom de la sous-catégorie'");
-      return false;
-    }
-
-    if (!newSubcategoryCode) {
-      notifyError("Veuillez remplir le champ 'Code de la sous-catégorie'");
-      return false;
-    }
-    if (row.subcategory.find((sub) => sub.label === newSubcategoryLabel)) {
-      notifyError("Cette sous-catégorie existe déjà");
-      return false;
-    }
-    return true;
-  };
-
-  const handleClick = async () => {
-    if (!handleValidation()) return;
-
-    try {
-      await addASubcategory({
-        variables: {
-          label: newSubcategoryLabel,
-          code: newSubcategoryCode,
-          categoryId: row.categoryId,
-        },
-        refetchQueries: ["GetCategories"],
-      });
-
-      setNewSubcategoryLabel("");
-      setNewSubcategoryCode("");
-      setShowInput(!showInput);
-      notifySuccess("Sous-catégorie ajoutée avec succès");
-    } catch (err) {
-      console.info(err);
-      notifyError("Erreur lors de l'ajout de la sous-catégorie");
-    }
-  };
 
   interface Subcategory {
     id: number;
@@ -444,110 +393,10 @@ function Row(props: {
                     </TableRow>
                   ))}
 
-                  <TableRow>
-                    <TableCell
-                      sx={{ paddingBottom: "5vh", fontSize: "1.1rem" }}
-                    >
-                      <Typography>
-                        <IconButton onClick={() => setShowInput(!showInput)}>
-                          <AddCircleOutline />
-                        </IconButton>
-                        Ajouter une sous-catégorie
-                      </Typography>
-                      {showInput && (
-                        <Box
-                          sx={{
-                            marginTop: "2rem",
-                            padding: "1rem",
-                            backgroundColor: "white",
-                            borderRadius: "8px",
-                            boxShadow: "0px 4px 6px rgba(0, 0, 0, 0.1)",
-                          }}
-                        >
-                          <Box
-                            sx={{
-                              display: "flex",
-                              flexDirection: "column",
-                              gap: "1rem",
-                            }}
-                          >
-                            <Box
-                              component="input"
-                              type="text"
-                              placeholder="Nom de la sous-catégorie"
-                              sx={{
-                                width: "100%",
-                                maxWidth: "300px",
-                                height: "4vh",
-                                fontSize: "1.1rem",
-                                textAlign: "center",
-                                padding: "0.5vh",
-                              }}
-                              value={newSubcategoryLabel}
-                              onChange={(e) =>
-                                setNewSubcategoryLabel(e.target.value)
-                              }
-                            />
-                            <Box
-                              component="input"
-                              type="text"
-                              placeholder="Code de la sous-catégorie"
-                              sx={{
-                                width: "100%",
-                                maxWidth: "300px",
-                                height: "4vh",
-                                fontSize: "1.1rem",
-                                textAlign: "center",
-                                padding: "0.5vh",
-                              }}
-                              value={newSubcategoryCode}
-                              onChange={(e) =>
-                                setNewSubcategoryCode(e.target.value)
-                              }
-                            />
-                            <Box
-                              sx={{
-                                display: "flex",
-                                gap: "1rem",
-                                justifyContent: "space-between",
-                              }}
-                            >
-                              <Button
-                                sx={{
-                                  width: "100%",
-                                  maxWidth: "120px",
-                                  height: "4vh",
-                                  fontSize: "1.1rem",
-                                  backgroundColor: theme.palette.success.main,
-                                  color: "white",
-                                  border: "none",
-                                  fontWeight: "bold",
-                                }}
-                                onClick={handleClick}
-                              >
-                                Valider
-                              </Button>
-                              <Button
-                                sx={{
-                                  width: "100%",
-                                  maxWidth: "120px",
-                                  height: "4vh",
-                                  fontSize: "1.1rem",
-                                  backgroundColor: theme.palette.error.main,
-                                  color: "white",
-                                  border: "none",
-                                  fontWeight: "bold",
-                                }}
-                                onClick={() => setShowInput(false)}
-                              >
-                                Annuler
-                              </Button>
-                            </Box>
-                          </Box>
-                        </Box>
-                      )}
-                    </TableCell>
-                  </TableRow>
+                  <AddSubcategory
+                    categoryId={row.categoryId}
+                    subcategories={row.subcategory}
+                  />
                 </TableBody>
               </Table>
             </Box>
